requirements.txt
streamlit
chromadb
sentence-transformers
groq
python-dotenv
duckduckgo-search
pandas
